refactor: moved functions deleting and changing in TaskItem and minor fixes
